requirements.txt
aiogram
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic
python-dotenv
jinja2
python-multipart
Pillow
